adds a check to powerloader grabbing

+++ code/modules/vehicles/powerloader.dm
@@ -304,17 +304,19 @@
 
 	if(!load_target_tag)
 		return
-	grab_object(target, load_target_tag)
-	user.visible_message(SPAN_NOTICE("[user] grabs \the [loaded] with \the [src]."),
-	SPAN_NOTICE("You grab \the [loaded] with \the [src]."))
+	grab_object(user, target, load_target_tag)
 
 //a bit unsafe proc
-/obj/item/powerloader_clamp/proc/grab_object(var/obj/target, var/target_tag = "", var/sound = 'sound/machines/hydraulics_2.ogg')
+/obj/item/powerloader_clamp/proc/grab_object(var/mob/user, var/obj/target, var/target_tag = "", var/sound = 'sound/machines/hydraulics_2.ogg')
+	if(loaded)
+		to_chat(user, SPAN_WARNING("\The [src] must be empty in order to grab \the [target]!"))
+		return
 	if(!linked_powerloader)
 		qdel(src)
 		return
 	loaded = target
 	loaded.forceMove(src)
+	to_chat(user, SPAN_NOTICE("You grab \the [target] with \the [src]."))
 	playsound(src, sound, 40, 1)
 	update_icon(target_tag)
 	target.update_icon()
